[type]: #75 separate PASS command implementation into own file
[body]

[notes]

close #75
